requirements.txt
viam-sdk~=0.47.0
pyinstaller~=6.10.0
